+++ klippy/extras/force_move.py
@@ -117,7 +117,7 @@ class ForceMove:
             accel,
         )
         print_time = print_time + accel_t + cruise_t + accel_t
-        toolhead.note_mcu_movequeue_activity(print_time)
+        self.motion_queuing.note_mcu_movequeue_activity(print_time)
         toolhead.dwell(accel_t + cruise_t + accel_t)
         toolhead.flush_step_generation()
         stepper.set_trapq(prev_trapq)
