+++ klippy/extras/manual_stepper.py
@@ -102,8 +102,7 @@ class ManualStepper:
         self.next_cmd_time = self._submit_move(
             self.next_cmd_time, movepos, speed, accel
         )
-        toolhead = self.printer.lookup_object("toolhead")
-        toolhead.note_mcu_movequeue_activity(self.next_cmd_time)
+        self.motion_queuing.note_mcu_movequeue_activity(self.next_cmd_time)
         if sync:
             self.sync_print_time()
 
@@ -266,8 +265,9 @@ class ManualStepper:
             start_time, newpos[0], speed, self.homing_accel
         )
         # Drip updates to motors
-        toolhead = self.printer.lookup_object("toolhead")
-        toolhead.drip_update_time(start_time, end_time, drip_completion)
+        self.motion_queuing.drip_update_time(
+            start_time, end_time, drip_completion
+        )
         # Clear trapq of any remaining parts of movement
         reactor = self.printer.get_reactor()
         self.motion_queuing.wipe_trapq(self.trapq)
